requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.4.0
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
whitenoise==4.1.2
